server and client error handling

+++ src/main/java/com/manalrzzl/professionalproject/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.manalrzzl.professionalproject.student;
 
+import com.manalrzzl.professionalproject.student.exception.BadRequestException;
+import com.manalrzzl.professionalproject.student.exception.NotFoundException;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,11 +17,17 @@ public class StudentService {
   }
 
   public void addStudent(Student student) {
-    // check if email is taken
+    //TODO:check if email is taken
+    if (studentRepository.selectExistsEmail(student.getEmail())) {
+      throw new BadRequestException("Email already taken");
+    }
     studentRepository.save(student);
   }
 
   public void deleteStudent(Long id) {
+    if(!studentRepository.existsById(id)) {
+      throw new NotFoundException("Student not found");
+    }
     studentRepository.deleteById(id);
   }
 }
